third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles/styles-do-not-add-inline-stylesheets-in-navigator.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<style>
</style>
<script>
function injectStyleSheet()
{
    var style = document.createElement('style');
    style.textContent = '* {color: blue; }';
    document.head.appendChild(style);
}

function test() {
  Promise.all([UI.inspectorView.showPanel('sources'), TestRunner.evaluateInPagePromise('injectStyleSheet()')])
      .then(onInjected);

  function onInjected() {
    var sourcesNavigator = new Sources.SourcesNavigatorView();
    SourcesTestRunner.dumpNavigatorView(sourcesNavigator);
    TestRunner.completeTest();
  }
}
</script>
</head>

<body onload="runTest()">
<p>Verify that inline stylesheets do not appear in navigator.</p>
</body>
</html>
